Enable global mode in use-package example
At least in my setup (spacemacs 0.300 on emacs 26.1), I could not reproduce the effects of the config without use-package if I just picked the use-package alternative. Until I noticed that they are not exactly the same.

Adding it to the example so that others do not get confused with the same mistake.